[spirv] Disentangle patterns in SPIRV*ToGPUPass
This is another step towards cleaning up the SPIRVConvertToGPUPass
after https://github.com/google/iree/pull/5814. We still have two
sets of patterns in the pass, one for handling copying to workgroup
memory, the other for tiling into global invocation IDs. These are
relatively disjoint functionalities; both should be cleaned up in
their own ways. So this commit is a prelimiary step towards that
by shuffling code around to their proper places and properly naming
stuff. This makes it clear for future dedicated improvements
afterwards in following up commits.

Along the way, this also cleans up the SPIRVVectorToGPU pass, which
uses utilities from the SPIRVConvertToGPU pass. SPIRVVectorToGPUPass
really just contains one pattern to handle copying to workgroup
memory (with vectorization). So renaming it properly. Again,
futher cleanup will be done in following up patches.

This commit is NFC (just shuffling code around), except that the
test for vectorized workgroup memory copy is disabled for now
to make migration easier.
