+++ iree/compiler/Codegen/SPIRV/BUILD
@@ -20,10 +20,10 @@ cc_library(
         "Passes.cpp",
         "SPIRVConcretizeWorkgroupTiles.cpp",
         "SPIRVConvertToGPU.cpp",
+        "SPIRVCopyToWorkgroupMemory.cpp",
         "SPIRVFoldGPUProcessorIDUses.cpp",
         "SPIRVTileAndVectorize.cpp",
         "SPIRVVectorToCooperativeMatrix.cpp",
-        "SPIRVVectorToGPU.cpp",
         "SPIRVVectorizeLoadStore.cpp",
         "Utils.cpp",
     ],
