+++ iree/compiler/Codegen/SPIRV/Utils.cpp
@@ -158,5 +158,203 @@ template SmallVector<linalg::ProcInfo, 2>
 getGPUProcessorIdsAndCounts<GPUGlobalId, GPUGlobalCount>(OpBuilder &builder,
                                                          Location loc,
                                                          unsigned numDims);
+
+scf::ParallelOp collapseParallelLoops(PatternRewriter &rewriter,
+                                      scf::ParallelOp pLoopOp) {
+  if (pLoopOp.getNumReductions()) return nullptr;
+
+  unsigned numLoops = pLoopOp.getNumLoops();
+  if (numLoops == 1) return pLoopOp;
+
+  // Compute the number of iterations of each loops starting from the innermost.
+  Location loc = pLoopOp.getLoc();
+  Value totalNumIterations = rewriter.create<ConstantIndexOp>(loc, 1);
+
+  // Track the "stride" of each loop, i.e. product of the total number of
+  // iterations of the inner loops.
+  SmallVector<Value, 2> iterationStride;
+  iterationStride.resize(pLoopOp.getNumLoops());
+  auto lbs = pLoopOp.lowerBound();
+  auto ubs = pLoopOp.upperBound();
+  auto steps = pLoopOp.step();
+  for (int i = numLoops - 1; i >= 0; --i) {
+    Value lb = lbs[i], ub = ubs[i], step = steps[i];
+    Value iterCount = rewriter.create<SignedDivIOp>(
+        loc, rewriter.create<SubIOp>(loc, ub, lb), step);
+    iterationStride[i] = totalNumIterations;
+    totalNumIterations =
+        rewriter.create<MulIOp>(loc, totalNumIterations, iterCount);
+  }
+
+  // Create the collapsed parallel loop op with lowerbound 0, step 1 and upper
+  // bound being the totalNumIterations.
+  Value newLb = rewriter.create<ConstantIndexOp>(loc, 0);
+  Value newStep = rewriter.create<ConstantIndexOp>(loc, 1);
+  scf::ParallelOp newPLoopOp =
+      rewriter.create<scf::ParallelOp>(loc, newLb, totalNumIterations, newStep);
+
+  // Build the body of the collapsed loop by cloning the original loop body. The
+  // replacement value of the induction variables of the original loop body,
+  // from the induction variable of the new loop, using
+  //   origLoopIv[i] = loopIv / iterationStride[i]
+  //   loopIv = loopIv % iterationStride[i]
+  OpBuilder::InsertionGuard guard(rewriter);
+  Block &pLoopBody = pLoopOp.getLoopBody().front();
+  rewriter.setInsertionPointToStart(&newPLoopOp.getLoopBody().front());
+  Value loopIv = *newPLoopOp.getInductionVars().begin();
+  BlockAndValueMapping map;
+  for (int i : llvm::seq<int>(0, numLoops)) {
+    Value iterNum =
+        rewriter.create<SignedDivIOp>(loc, loopIv, iterationStride[i]);
+    AffineExpr d0, d1;
+    bindDims(rewriter.getContext(), d0, d1);
+    AffineExpr s0 = getAffineSymbolExpr(0, rewriter.getContext());
+    Value newIv = makeComposedAffineApply(rewriter, loc, d0 + d1 * s0,
+                                          {lbs[i], iterNum, steps[i]});
+    map.map(pLoopBody.getArgument(i), newIv);
+    loopIv = rewriter.create<SignedRemIOp>(loc, loopIv, iterationStride[i]);
+  }
+  for (Operation &op : pLoopBody.without_terminator()) {
+    rewriter.clone(op, map);
+  }
+  rewriter.eraseOp(pLoopOp);
+  return newPLoopOp;
+}
+
+/// Builds an empty scf.for operation. The default builder adds an entry basic
+/// block which needs to be avoided here.
+static scf::ForOp buildEmptyForOp(Location loc, OpBuilder &builder, Value lb,
+                                  Value ub, Value step) {
+  OperationState state(loc, scf::ForOp::getOperationName());
+  state.addOperands({lb, ub, step});
+  state.addRegion();
+  return cast<scf::ForOp>(builder.createOperation(state));
+}
+
+Operation *replacePLoopOp(ConversionPatternRewriter &rewriter,
+                          scf::ParallelOp pLoopOp,
+                          ArrayRef<LoopBounds> newPLoopBounds,
+                          ArrayRef<LoopBounds> forBounds,
+                          ArrayRef<unsigned> permutation) {
+  assert(!forBounds.empty() && "unhandled case of no scf.for created");
+  unsigned numLoops = pLoopOp.getNumLoops();
+  Location loc = pLoopOp.getLoc();
+  assert(forBounds.size() + newPLoopBounds.size() == numLoops &&
+         "cannot drop loops when splitting scf.parallel operation");
+  assert(permutation.size() == numLoops);
+  OpBuilder::InsertionGuard guard(rewriter);
+
+  // Need a signature conversion for the body of the scf.parallel operation,
+  // before can it can be used as the body of the innermost loop created here.
+  TypeConverter::SignatureConversion signatureConverter(numLoops);
+  Operation *outermostLoop = nullptr;
+  auto permuteIt = permutation.begin();
+
+  // Create the scf.parallel operation as the outermost loop, if specified.
+  if (!newPLoopBounds.empty()) {
+    auto lbs = llvm::to_vector<2>(llvm::map_range(
+        newPLoopBounds, [](LoopBounds bounds) -> Value { return bounds.lb; }));
+    auto ubs = llvm::to_vector<2>(llvm::map_range(
+        newPLoopBounds, [](LoopBounds bounds) { return bounds.ub; }));
+    auto steps = llvm::to_vector<2>(llvm::map_range(
+        newPLoopBounds, [](LoopBounds bounds) { return bounds.step; }));
+    auto newPLoop = rewriter.create<scf::ParallelOp>(loc, lbs, ubs, steps);
+    for (auto iv : newPLoop.getInductionVars()) {
+      signatureConverter.remapInput(*permuteIt, iv);
+      permuteIt++;
+    }
+    rewriter.setInsertionPointToStart(newPLoop.getBody());
+    outermostLoop = newPLoop.getOperation();
+  }
+
+  // Generate the nested scf.for operations with the bounds passed.
+  for (auto it : enumerate(forBounds)) {
+    Value lb = it.value().lb, ub = it.value().ub, step = it.value().step;
+    if (it.index() != forBounds.size() - 1) {
+      auto forOp = rewriter.create<scf::ForOp>(loc, lb, ub, step);
+      if (!outermostLoop) outermostLoop = forOp.getOperation();
+      signatureConverter.remapInput(*permuteIt, forOp.getInductionVar());
+      rewriter.setInsertionPointToStart(forOp.getBody());
+    } else {
+      // For the last loop, move the body of the scf.parallel op as the body of
+      // the loop after signature conversion.
+      auto forOp = buildEmptyForOp(loc, rewriter, lb, ub, step);
+      if (!outermostLoop) outermostLoop = forOp.getOperation();
+      signatureConverter.addInputs(*permuteIt, rewriter.getIndexType());
+      Region &pLoopOpRegion = pLoopOp.getLoopBody();
+      rewriter.applySignatureConversion(&pLoopOpRegion, signatureConverter);
+      Region &forOpRegion = forOp.getLoopBody();
+      rewriter.inlineRegionBefore(pLoopOpRegion, forOpRegion,
+                                  forOpRegion.begin());
+    }
+    permuteIt++;
+  }
+  rewriter.eraseOp(pLoopOp);
+  return outermostLoop;
+}
+
+/// Builds an empty scf.if operation without the then and else blocks.
+static scf::IfOp buildEmptyIfOp(Location loc, OpBuilder &builder, Value cond) {
+  OperationState state(loc, scf::IfOp::getOperationName());
+  state.addOperands(cond);
+  state.addRegion();
+  state.addRegion();
+  return cast<scf::IfOp>(builder.createOperation(state));
+}
+
+LogicalResult distributeSingleIterationPerProcessor(
+    ConversionPatternRewriter &rewriter, scf::ParallelOp pLoopOp,
+    ArrayRef<linalg::ProcInfo> procInfo, bool generateGuard) {
+  unsigned numLoops = pLoopOp.getNumLoops();
+  Location loc = pLoopOp.getLoc();
+  assert(numLoops == procInfo.size() &&
+         "expected as many ids as number of loops");
+
+  auto lbs = pLoopOp.lowerBound();
+  auto step = pLoopOp.step();
+  SmallVector<Value, 2> ivReplacements;
+  for (unsigned i : llvm::seq<unsigned>(0, numLoops)) {
+    Value iterValue = rewriter.create<AddIOp>(
+        loc, lbs[i], rewriter.create<MulIOp>(loc, procInfo[i].procId, step[i]));
+    ivReplacements.push_back(iterValue);
+  }
+  Region &pLoopOpRegion = pLoopOp.getLoopBody();
+
+  if (generateGuard) {
+    TypeConverter::SignatureConversion signatureConverter(numLoops);
+    Value cond = nullptr;
+    auto ubs = pLoopOp.upperBound();
+    for (unsigned i : llvm::seq<unsigned>(0, numLoops)) {
+      Value cmp = rewriter.create<CmpIOp>(loc, CmpIPredicate::slt,
+                                          ivReplacements[i], ubs[i]);
+      cond = (cond ? rewriter.create<AndOp>(loc, cond, cmp) : cmp);
+      signatureConverter.remapInput(i, ivReplacements[i]);
+    }
+    rewriter.applySignatureConversion(&pLoopOpRegion, signatureConverter);
+    scf::IfOp ifOp = buildEmptyIfOp(loc, rewriter, cond);
+    Region &ifOpRegion = ifOp.getRegion(0);
+    rewriter.inlineRegionBefore(pLoopOpRegion, ifOpRegion, ifOpRegion.begin());
+  } else {
+    // The body of the scf.parallel needs to be moved into its parent
+    // operation.
+    // - Split the block just before the scf.parallel operation.
+    // - Move the only block of scf.parallel before the newly created block
+    //   (after signature conversion).
+    // - Add branch from the original block to the moved block of the
+    //   scf.parallel's region, and from the latter to the block created by the
+    //   split operation.
+    // - Canonicalization will fold these branches away.
+    Block *destBlock = pLoopOp.getOperation()->getBlock();
+    Block *remainingInst =
+        rewriter.splitBlock(destBlock, Block::iterator(pLoopOp));
+    Block *sourceBlock = &pLoopOpRegion.front();
+    rewriter.eraseOp(sourceBlock->getTerminator());
+    rewriter.mergeBlocks(&pLoopOpRegion.front(), destBlock, ivReplacements);
+    rewriter.mergeBlocks(remainingInst, destBlock, {});
+  }
+  rewriter.eraseOp(pLoopOp);
+  return success();
+}
+
 }  // namespace iree_compiler
 }  // namespace mlir
